Moved the L10n to its own file
